Move graph traversal functions into their own file
