+++ pytensor/ifelse.py
@@ -21,10 +21,11 @@ import pytensor.tensor as pt
 from pytensor import as_symbolic
 from pytensor.compile import optdb
 from pytensor.configdefaults import config
-from pytensor.graph.basic import Apply, Variable, apply_depends_on
+from pytensor.graph.basic import Apply, Variable
 from pytensor.graph.op import _NoPythonOp
 from pytensor.graph.replace import clone_replace
 from pytensor.graph.rewriting.basic import GraphRewriter, in2out, node_rewriter
+from pytensor.graph.traversal import apply_depends_on
 from pytensor.graph.type import HasDataType, HasShape
 from pytensor.tensor.shape import Reshape, Shape, SpecifyShape
