+++ pytensor/scan/utils.py
@@ -15,9 +15,10 @@ from pytensor import scalar as ps
 from pytensor import tensor as pt
 from pytensor.compile.profiling import ProfileStats
 from pytensor.configdefaults import config
-from pytensor.graph.basic import Constant, Variable, equal_computations, graph_inputs
+from pytensor.graph.basic import Constant, Variable, equal_computations
 from pytensor.graph.op import get_test_value
 from pytensor.graph.replace import clone_replace
+from pytensor.graph.traversal import graph_inputs
 from pytensor.graph.type import HasDataType
 from pytensor.graph.utils import TestValueError
 from pytensor.tensor.basic import AllocEmpty, cast
