+++ tests/tensor/test_variable.py
@@ -10,6 +10,7 @@ import tests.unittest_tools as utt
 from pytensor.compile import DeepCopyOp
 from pytensor.compile.mode import get_default_mode
 from pytensor.graph.basic import Constant, equal_computations
+from pytensor.graph.traversal import io_toposort
 from pytensor.tensor import get_vector_length
 from pytensor.tensor.basic import constant
 from pytensor.tensor.elemwise import DimShuffle
@@ -144,7 +145,7 @@ def test__getitem__Subtensor():
     i = iscalar("i")
 
     z = x[i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == Subtensor
 
     # This should ultimately do nothing (i.e. just return `x`)
@@ -156,29 +157,29 @@ def test__getitem__Subtensor():
     # It lands in the `full_slices` condition in
     # `_tensor_py_operators.__getitem__`
     z = x[..., None]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert all(op_type == DimShuffle for op_type in op_types)
 
     z = x[None, :, None, :]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert all(op_type == DimShuffle for op_type in op_types)
 
     # This one lands in the non-`full_slices` condition in
     # `_tensor_py_operators.__getitem__`
     z = x[:i, :, None]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[1:] == [DimShuffle, Subtensor]
 
     z = x[:]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == Subtensor
 
     z = x[..., :]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == Subtensor
 
     z = x[..., i, :]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == Subtensor
 
 
@@ -187,24 +188,24 @@ def test__getitem__AdvancedSubtensor_bool():
     i = TensorType("bool", shape=(None, None))("i")
 
     z = x[i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == AdvancedSubtensor
 
     i = TensorType("bool", shape=(None,))("i")
     z = x[:, i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == AdvancedSubtensor
 
     i = TensorType("bool", shape=(None,))("i")
     z = x[..., i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == AdvancedSubtensor
 
     with pytest.raises(TypeError):
         z = x[[True, False], i]
 
     z = x[ivector("b"), i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == AdvancedSubtensor
 
 
@@ -215,26 +216,26 @@ def test__getitem__AdvancedSubtensor():
 
     # This is a `__getitem__` call that's redirected to `_tensor_py_operators.take`
     z = x[i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == AdvancedSubtensor
 
     # This should index nothing (i.e. return an empty copy of `x`)
     # We check that the index is empty
     z = x[[]]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types == [AdvancedSubtensor]
     assert isinstance(z.owner.inputs[1], TensorConstant)
 
     z = x[:, i]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types == [MakeSlice, AdvancedSubtensor]
 
     z = x[..., i, None]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types == [MakeSlice, AdvancedSubtensor]
 
     z = x[i, None]
-    op_types = [type(node.op) for node in pytensor.graph.basic.io_toposort([x, i], [z])]
+    op_types = [type(node.op) for node in io_toposort([x, i], [z])]
     assert op_types[-1] == AdvancedSubtensor
